Extracted the AddNewNote to a separated class.
